refactor: align example naming
