+++ crd-generator/src/test/java/io/dekorate/crd/example/person/Address.java
@@ -1,28 +1,23 @@
 /**
  * Copyright 2018 The original authors.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *     http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- * 
-**/
+ */
+package io.dekorate.crd.example.person;
 
-package io.dekorate.crd.examples.webserver;
-
-import io.dekorate.crd.alt.SpecReplicas;
-
-public class WebServerSpec {
-
-  private int port;
-
-  @SpecReplicas
-  private int replicas;
+public class Address {
+  private String street;
+  private int number;
+  private String zip;
+  private String country;
 }
