+++ crd-generator/src/test/java/io/dekorate/crd/util/JsonSchemaTest.java
@@ -19,7 +19,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
 
-import io.dekorate.crd.examples.person.Person;
+import io.dekorate.crd.example.person.Person;
 import io.dekorate.utils.Serialization;
 import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.JSONSchemaProps;
 import io.sundr.codegen.functions.ClassTo;
